Change from int to long

+++ src/Main.java
@@ -3,7 +3,7 @@ import java.util.Arrays;
 public class Main {
 
     public static void main(String[] args) {
-        var SalesManager = new SalesManager(new int[]{15, 27, 30, 42, 54});
+        var SalesManager = new SalesManager(new long[]{15, 27, 30, 42, 54});
         System.out.printf("Максимум от {15, 27, 30, 42, 54} = %d\n", SalesManager.max());
     }
 }
